Fix larproperties.fcl
Current FastScintEnergies and SlowScintEnergies seem to correspond to Kr (emission peak at ~8.5 eV -> 146 nm) instead of Argon (emission peak at ~9.7 eV -> 128 nm). There is no need to change the shape of the spectra (FastScintSpectrum/SlowScintSpectrum) but energies need to be shifted in ~1.2 eV to larger values (which is the proposed change).